Degree Id added in get active stream and negative marks updated to de…

+++ APTITUDE/APTITUDETEST.Core/ViewModels/ResultsVM.cs
@@ -10,7 +10,7 @@
         public decimal Points { get; set; }
         public int CorrectMarks { get; set; }
         public int CorrectCount { get; set; }
-        public int WrongMarks { get; set; }
+        public decimal WrongMarks { get; set; }
         public int WrongCount { get; set; }
         public int UnAnsweredCount { get; set; }
         public int UnDisplayedCount { get; set; }
